Rework tests structure
