+++ tests/routedmessagetests.pro
@@ -1,0 +1,8 @@
+QT += widgets testlib
+QT -= gui
+TARGET = routedmessagetests
+DEFINES += QMQTT_LIBRARY_TESTS
+CONFIG += testcase
+SOURCES += routedmessagetests.cpp
+INCLUDEPATH += ..
+LIBS += -L.. -lqmqtt
