+++ tests/tests.pri
@@ -1,36 +1,19 @@
-TEST_SOURCES += \
-#    testmain.cpp \
-#    willtests.cpp \
-#    messagetests.cpp \
-#    frametests.cpp \
-    networktests.cpp \
-    tcpserver.cpp \
-#    routedmessagetests.cpp \
-#    routertests.cpp \
-#    routesubscriptiontests.cpp \
-    clienttests.cpp
+OTHER_FILES += \
+    $${TEST_FILES} \
+    tests/tests.pri
 
-TEST_HEADERS += \
-#    willtests.h \
-#    messagetests.h \
-#    frametests.h \
-#    networktests.h \
-    tcpserver.h \
-#    routedmessagetests.h \
-#    routertests.h \
-#    routesubscriptiontests.h
-
-for(file, TEST_SOURCES) {
-     TEST_FILES += tests/$${file}
+unix {
+    unit_tests_directory = tests
+    unit_tests.target = all
+    unit_tests.commands = \
+        mkdir -p $${OUT_PWD}/$${unit_tests_directory}; \
+        cd $${OUT_PWD}/$${unit_tests_directory}; \
+        for pro_path in $$find(TEST_FILES,\.pro$); do \
+            pro_filename=\$\${pro_path$${LITERAL_HASH}$${LITERAL_HASH}*/}; \
+            pro_basename=\$\${pro_filename%.pro}; \
+            $${QMAKE_QMAKE} -o Makefile.\$\${pro_basename} $${PWD}/\$\${pro_filename}; \
+            LD_LIBRARY_PATH=$${OUT_PWD} make -f Makefile.\$\${pro_basename} check; \
+        done; \
+        cd $${OUT_PWD}
+    QMAKE_EXTRA_TARGETS += unit_tests
 }
-
-for(file, TEST_HEADERS) {
-     TEST_FILES += tests/$${file}
-}
-
-TEST_FILES += \
-#    tests/tests.pro \
-    tests/test.pri
-
-OTHER_FILES += $${TEST_FILES}
-
